Change grid structure
Instead of having one column with the lists on the left side and one
column with searching and dates on the right side, this is added as a
row instead.

It is possible that these options could be integrated into the header
in order to make the UI even better.